[STEP1] TempUnit 수정 	- 그룹화
