+++ WeatherForecast/Foundation/TempUnit/Protocols/TempUnit.swift
@@ -12,12 +12,3 @@ protocol TempUnit {
     var expression: String { get }
     var expressionTitle: String { get }
 }
-struct Metric: TempUnit {
-    var expression: String = "℉"
-    var expressionTitle: String = "화씨"
-}
-
-struct Imperial: TempUnit {
-    var expression: String = "℃"
-    var expressionTitle: String = "섭씨"
-}
