+++ WeatherForecast/Foundation/TempUnit/Structs/Imperial.swift
@@ -1,0 +1,13 @@
+//
+//  Imperial.swift
+//  WeatherForecast
+//
+//  Created by 김창규 on 1/31/24.
+//
+
+import Foundation
+
+struct Imperial: TempUnit {
+    var expression: String = "℃"
+    var expressionTitle: String = "섭씨"
+}
