+++ WeatherForecast/Foundation/TempUnit/Structs/Metric.swift
@@ -1,0 +1,13 @@
+//
+//  Metric.swift
+//  WeatherForecast
+//
+//  Created by 김창규 on 1/31/24.
+//
+
+import Foundation
+
+struct Metric: TempUnit {
+    var expression: String = "℉"
+    var expressionTitle: String = "화씨"
+}
